Update Pseudocode
Added some to the list of required assets and expected organization of the network library. Some of this is admittedly a little too much of a change from the previous iteration, so feel free to tear it apart in the case that it falls short in some respects. I'll look it over again and add/update.

This is just a bulleted list at the moment. Pseudocode is forthcoming, since there are plenty of algorithms to coordinate with this project, and I think having this reference should prove useful.